Move rotation/angle functions to a different file
